refactoring language selector
