+++ src/main/java/de/fred4jupiter/fredbet/web/language/LanguageSupport.java
@@ -1,0 +1,8 @@
+package de.fred4jupiter.fredbet.web.language;
+
+public record LanguageSupport(String languageCode, String flagCode) {
+
+    public LanguageSupport(String languageCode) {
+        this(languageCode, languageCode);
+    }
+}
