+++ src/main/resources/templates/fragments/language_selector.html
@@ -2,64 +2,10 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <body>
 	<ul th:fragment="language_selector" th:remove="tag">
-        <li>
-            <a th:href="${@languageUtil.buildUrl('de')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('de')}"></span>
-                <span th:text="' '+#{language.de}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('en')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('gb')}"></span>
-                <span th:text="' '+#{language.en}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('nl')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('nl')}"></span>
-                <span th:text="' '+#{language.nl}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('fr')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('fr')}"></span>
-                <span th:text="' '+#{language.fr}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('es')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('es')}"></span>
-                <span th:text="' '+#{language.es}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('ca')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('es-ct')}"></span>
-                <span th:text="' '+#{language.ca}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('pl')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('pl')}"></span>
-                <span th:text="' '+#{language.pl}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('cs')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('cz')}"></span>
-                <span th:text="' '+#{language.cs}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('sv')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('se')}"></span>
-                <span th:text="' '+#{language.sv}"></span>
-            </a>
-        </li>
-        <li>
-            <a th:href="${@languageUtil.buildUrl('et')}" class="dropdown-item language">
-                <span class="fi" th:classappend="${@languageUtil.cssClassFor('ee')}"></span>
-                <span th:text="' '+#{language.et}"></span>
+        <li th:each="langSupport : ${@languageUtil.SUPPORTED_LANGUAGES}">
+            <a th:href="${@languageUtil.buildUrl(langSupport.languageCode())}" class="dropdown-item language">
+                <span class="fi" th:classappend="${@languageUtil.cssClassFor(langSupport.flagCode())}"></span>
+                <span th:text="' '+#{'language.' + ${langSupport.languageCode()}}"></span>
             </a>
         </li>
 	</ul>
