+++ src/test/java/de/fred4jupiter/fredbet/web/WebUtilitiesUT.java
@@ -3,6 +3,7 @@ package de.fred4jupiter.fredbet.web;
 import de.fred4jupiter.fredbet.common.UnitTest;
 import de.fred4jupiter.fredbet.domain.Group;
 import de.fred4jupiter.fredbet.match.MatchService;
+import de.fred4jupiter.fredbet.web.language.LanguageUtil;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.springframework.context.i18n.LocaleContextHolder;
